+++ http4k-core/src/main/kotlin/org/http4k/client/JavaHttpClient.kt
@@ -1,61 +1,75 @@
 package org.http4k.client
 
+import org.http4k.client.PreCannedJavaHttpClients.defaultJavaHttpClient
 import org.http4k.core.Body
+import org.http4k.core.BodyMode
+import org.http4k.core.BodyMode.Memory
+import org.http4k.core.BodyMode.Stream
 import org.http4k.core.HttpHandler
 import org.http4k.core.Request
 import org.http4k.core.Response
 import org.http4k.core.Status
 import org.http4k.core.Status.Companion.CONNECTION_REFUSED
 import org.http4k.core.Status.Companion.UNKNOWN_HOST
-import java.io.ByteArrayInputStream
 import java.net.ConnectException
-import java.net.HttpURLConnection
-import java.net.URL
+import java.net.URI
 import java.net.UnknownHostException
+import java.net.http.HttpClient
+import java.net.http.HttpClient.Redirect.NEVER
+import java.net.http.HttpClient.Version.HTTP_1_1
+import java.net.http.HttpRequest
+import java.net.http.HttpResponse
+import java.net.http.HttpResponse.BodyHandlers
 import java.nio.ByteBuffer
 
-object JavaHttpClient {
-    operator fun invoke(): HttpHandler = { request: Request ->
-        try {
-            val connection = (URL(request.uri.toString()).openConnection() as HttpURLConnection).apply {
-                instanceFollowRedirects = false
-                requestMethod = request.method.name
-                doOutput = true
-                doInput = true
-                request.headers.forEach {
-                    addRequestProperty(it.first, it.second)
-                }
-                request.body.apply {
-                    if (this != Body.EMPTY) {
-                        val content = if (stream.available() == 0) payload.array().inputStream() else stream
-                        content.copyTo(outputStream)
-                    }
-                }
-            }
-
-            val status = Status(connection.responseCode, connection.responseMessage.orEmpty())
-            val baseResponse = Response(status).body(connection.body(status))
-            connection.headerFields
-                .filterKeys { it != null } // because response status line comes as a header with null key (*facepalm*)
-                .map { header -> header.value.map { header.key to it } }
-                .flatten()
-                .fold(baseResponse) { acc, next -> acc.header(next.first, next.second) }
-        } catch (e: UnknownHostException) {
-            Response(UNKNOWN_HOST.toClientStatus(e))
-        } catch (e: ConnectException) {
-            Response(CONNECTION_REFUSED.toClientStatus(e))
+/**
+ * Basic JDK-based Client.
+ */
+class JavaHttpClient(private val httpClient: HttpClient = defaultJavaHttpClient(),
+                     private val requestBodyMode: BodyMode = Memory,
+                     private val responseBodyMode: BodyMode = Memory
+) : HttpHandler {
+    override fun invoke(request: Request): Response = try {
+        val javaRequest = request.toJavaHttpRequest(requestBodyMode)
+        when (responseBodyMode) {
+            is Memory -> httpClient.send(javaRequest, BodyHandlers.ofByteArray())
+                .run { coreResponse().body(Body(ByteBuffer.wrap(body()))) }
+            is Stream -> httpClient.send(javaRequest, BodyHandlers.ofInputStream())
+                .run { coreResponse().body(body()) }
         }
+    } catch (e: UnknownHostException) {
+        Response(UNKNOWN_HOST.toClientStatus(e))
+    } catch (e: ConnectException) {
+        Response(CONNECTION_REFUSED.toClientStatus(e))
     }
+}
 
-    // Because HttpURLConnection closes the stream if a new request is made, we are forced to consume it straight away
-    private fun HttpURLConnection.body(status: Status) =
-        Body(resolveStream(status).readBytes().let { ByteBuffer.wrap(it) })
+object PreCannedJavaHttpClients {
+    /**
+     * Standard non-redirecting HTTP client
+     */
+    fun defaultJavaHttpClient(): HttpClient = HttpClient.newBuilder()
+        .version(HTTP_1_1)
+        .followRedirects(NEVER)
+        .build()
+}
 
-    private fun HttpURLConnection.resolveStream(status: Status) =
-        when {
-            status.serverError || status.clientError -> errorStream
-            else -> inputStream
-        } ?: EMPTY_STREAM
+private fun Request.toJavaHttpRequest(bodyMode: BodyMode) =
+    HttpRequest.newBuilder()
+        .uri(URI.create(uri.toString()))
+        .apply {
+            headers.fold(this) { acc, next -> acc.header(next.first, next.second) }
+        }.method(method.name, body.toRequestPublisher(bodyMode)).build()
 
-    private val EMPTY_STREAM = ByteArrayInputStream(ByteArray(0))
+private fun <T> HttpResponse<T>.coreResponse() =
+    Response(Status(statusCode(), "")).headers(headers().map()
+        .map { headerNameToValues ->
+            headerNameToValues.value
+                .map { headerNameToValues.key to it }
+        }
+        .flatten())
+
+private fun Body.toRequestPublisher(bodyMode: BodyMode) = when (bodyMode) {
+    Memory -> HttpRequest.BodyPublishers.ofByteArray(payload.array())
+    Stream -> HttpRequest.BodyPublishers.ofInputStream { stream }
 }